+++ Project.toml
@@ -5,8 +5,6 @@ version = "0.0.1-DEV"
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-CommonSolve = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 Ferrite = "c061ca5d-56c9-439f-9c0e-210fe06d3992"
@@ -42,7 +40,6 @@ CuThunderboltExt = "CUDA"
 [compat]
 Aqua = "0.8"
 BlockArrays = "1"
-CUDA = "5"
 DiffEqBase = "6"
 Ferrite = "1"
 LinearSolve = "2"
