+++ test/gpu/runtests.jl
@@ -1,8 +1,8 @@
-using Thunderbolt
 using CUDA
+using Thunderbolt
 using Test
 using StaticArrays
-
+x= 5
 
 include("test_operators.jl")
 include("test_coefficients.jl")
